adds debug that takes any type and stringifies!
